+++ data/yago.schema.sql
@@ -1,0 +1,5 @@
+create table triple (
+  subject varchar(100) not null,
+  predicate varchar(100) not null,
+  object varchar (100) not null
+);