+++ gen/yago/schema.h
@@ -1,0 +1,7 @@
+#ifndef _YAGO_SCHEMA_H_
+#define _YAGO_SCHEMA_H_
+
+#include "../../src/schema/store.h"
+#include "triple.h"
+
+#endif