+++ src/import.h
@@ -4,8 +4,10 @@
 #include <string>
 #include "tpcc.h"
 #include "tpce.h"
+#include "yago.h"
 
 void importSampleData(std::string, Tpcc*);
 void importSampleData(std::string, Tpce*);
+void importSampleData(std::string, Yago*);
 
 #endif