+++ src/yago.h
@@ -1,0 +1,14 @@
+#ifndef _YAGO_H_
+#define _YAGO_H_
+
+#include "../gen/yago/schema.h"
+
+/**
+ * Controller class for Yago tables.
+ */
+class Yago {
+  public:
+    yago::Triple facts;
+};
+
+#endif